+++ src/main/java/mod/azure/doom/item/weapons/Chaingun.java
@@ -73,7 +73,7 @@ public class Chaingun extends DoomBaseItem {
 	}
 
 	public static float getArrowVelocity(int charge) {
-		float f = (float) charge / 20.0F;
+		float f = charge / 20.0F;
 		f = (f * f + f * 2.0F) / 3.0F;
 		if (f > 1.0F) {
 			f = 1.0F;
@@ -115,7 +115,7 @@ public class Chaingun extends DoomBaseItem {
 	}
 
 	public static float getPullProgress(int useTicks) {
-		float f = (float) useTicks / 20.0F;
+		float f = useTicks / 20.0F;
 		f = (f * f + f * 2.0F) / 3.0F;
 		if (f > 1.0F) {
 			f = 1.0F;
