+++ src/main/java/mod/azure/doom/item/weapons/SuperShotgun.java
@@ -106,7 +106,7 @@ public class SuperShotgun extends DoomBaseItem {
 	}
 
 	public static float getArrowVelocity(int charge) {
-		float f = (float) charge / 20.0F;
+		float f = charge / 20.0F;
 		f = (f * f + f * 2.0F) / 3.0F;
 		if (f > 1.0F) {
 			f = 1.0F;
@@ -150,7 +150,7 @@ public class SuperShotgun extends DoomBaseItem {
 	}
 
 	public static float getPullProgress(int useTicks) {
-		float f = (float) useTicks / 20.0F;
+		float f = useTicks / 20.0F;
 		f = (f * f + f * 2.0F) / 3.0F;
 		if (f > 1.0F) {
 			f = 1.0F;
